requirements.txt
requests
chromadb
python-dotenv
tqdm
sentence-transformers